estrutura modificada para mais se assemelhar a logica de model-router-controller-service
