+++ models/curso-model.js
@@ -8,41 +8,34 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-const Sql = require("../infra/sql");
+const sql = require("../infra/sql");
 module.exports = class Curso {
     static validar(c) {
         return null;
     }
-    static listar() {
-        return __awaiter(this, void 0, void 0, function* () {
-            let lista = null;
-            yield Sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
-                lista = (yield sql.query("select idCurso, nomeCurso, responsavelCurso, horasSemanaisCurso, duracaoCurso, "
-                    + " descricaoCurso from curso order by nomeCurso asc"));
-            }));
-            return (lista || []);
-        });
-    }
-    // public static async obter(id: number): Promise<Contato> {
-    // 	let lista: Contato[] = null;
-    // 	await Sql.conectar(async (sql: Sql) => {
-    // 		lista = await sql.query("select id, nome, endereco, email, peso from contato where id = ?",[id]) as Contato[];
-    // 	});
-    // 	if (lista && lista[0]) {
-    // 		return lista[0];
-    // 	}else {
-    // 		return null;
-    // 	}
-    // 	//return ((lista && lista[0]) || null);
-    // }
     static criar(c) {
         return __awaiter(this, void 0, void 0, function* () {
             let res;
             if ((res = Curso.validar(c)))
                 return res;
-            yield Sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
+            yield sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
                 yield sql.query("insert into curso (nomeCurso,responsavelCurso,horasSemanaisCurso,duracaoCurso,descricaoCurso) values (?,?,?,?,?)", [c.nomeCurso, c.responsavelCurso, c.horasSemanaisCurso, c.duracaoCurso, c.descricaoCurso]);
             }));
+        });
+    }
+    static obter(id) {
+        return __awaiter(this, void 0, void 0, function* () {
+            let lista = null;
+            yield sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
+                lista = (yield sql.query("select idCurso, nomeCurso, responsavelCurso, horasSemanaisCurso, descricaoCurso, duracaoCursocurso where id = ?", [id]));
+            }));
+            if (lista && lista[0]) {
+                return lista[0];
+            }
+            else {
+                return null;
+            }
+            //return ((lista && lista[0]) || null);
         });
     }
     static alterar(c) {
@@ -50,7 +43,7 @@ module.exports = class Curso {
             let res;
             if ((res = Curso.validar(c)))
                 return res;
-            yield Sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
+            yield sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
                 yield sql.query("update curso set nomeCurso = ?, responsavelCurso = ?, horasSemanaisCurso = ?, duracaoCurso = ?, descricaoCurso = ? where idCurso = ?", [c.nomeCurso, c.responsavelCurso, c.horasSemanaisCurso, c.duracaoCurso, c.descricaoCurso, c.idCurso]);
                 res = sql.linhasAfetadas.toString();
             }));
@@ -59,12 +52,22 @@ module.exports = class Curso {
     static excluir(idCurso) {
         return __awaiter(this, void 0, void 0, function* () {
             let res = null;
-            yield Sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
+            yield sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
                 yield sql.query("delete from curso where idCurso = " + idCurso);
                 res = sql.linhasAfetadas.toString();
             }));
             return res;
         });
     }
+    static listar() {
+        return __awaiter(this, void 0, void 0, function* () {
+            let lista = null;
+            yield sql.conectar((sql) => __awaiter(this, void 0, void 0, function* () {
+                lista = (yield sql.query("select idCurso, nomeCurso, responsavelCurso, horasSemanaisCurso, duracaoCurso, "
+                    + " descricaoCurso from curso order by idCurso desc"));
+            }));
+            return (lista || []);
+        });
+    }
 };
-//# sourceMappingURL=curso.js.map
+//# sourceMappingURL=curso-model.js.map
